ai-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
branca==0.8.1
buildozer==1.5.0
cachetools==6.1.0
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
Django==5.2.4
docutils==0.21.2
dotenv==0.9.9
fastapi==0.135.1
filelock==3.16.1
filetype==1.2.0
Flask==3.1.2
folium==0.20.0
fonttools==4.59.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.49.1
google-genai==1.67.0
groq==0.37.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kafka==1.3.5
kafka-python==2.3.0
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
langchain==1.2.10
langchain-core==1.2.18
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.10
langsmith==0.7.16
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.6.3
narwhals==1.47.1
nltk==3.9.1
numpy==2.3.1
openai==1.93.2
opencage==3.2.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.1
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
phonenumbers==9.0.8
pillow==11.3.0
platformdirs==4.3.6
propcache==0.3.2
protobuf==6.31.1
ptyprocess==0.7.0
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.27.1
pyparsing==3.2.3
pypdfium2==5.6.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-whois==0.9.6
pytz==2025.2
pywin32==308
PyYAML==6.0.3
redis==7.1.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.10
requests==2.32.3
requests-file==3.0.1
requests-toolbelt==1.0.0
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.16.0
sh==2.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8.3
SpeechRecognition==3.14.3
sqlparse==0.5.3
srt==3.5.3
starlette==0.52.1
streamlit==1.47.0
streamlit-lottie==0.0.5
tavily-python==0.7.23
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tk==0.1.0
tldextract==5.3.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.3.0
uuid_utils==0.14.1
uv==0.7.13
uvicorn==0.41.0
virtualenv==20.28.1
vosk==0.3.45
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.5
xxhash==3.6.0
xyzservices==2025.4.0
yarl==1.20.1
zstandard==0.25.0
